refactor(AppliedRoute): move to Routing dir ♻️
